+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-
 ## [Unreleased]
 
 ### Added
 
+- Added `CONTRIBUTING.md` file with guidelines about contributing to the project
 - Added CI tests for the Spack package on Ubuntu 18.04, Debian 10, CentOS 7, Fedora 31, OpenSUSE Leap 15.2
 - Added `wget` and `autoconf` as buildtime dependencies in the Spack package
 - Added a documentation note about compiler selection when installing on CentOS 7 using the Spack package